Fix signature field eating remaining vars. All seems to be working
Also remove extra wait time at sign out. I think it was there for
debugging.

I don't like how the signing functionlity is abstracted right now,
either in `scope.setField.CANVAS` or `scope.sign`  -
it needs the handle of the canvas passed in. That makes sense for
other field setting, but right now signature pages always have
just one canvas that gets signed. That means everything that uses
those functions has to get the handle themselves. Because of the
way other things had to be done, it becomes unelegant. On the
other hand, I don't like deviating from the pattern of passing in
the handle. Not sure exactly why. Would love a second eye on this.