Fixe bugs of new loop. New problem - continuing on page without
fields then triggers continue button because 'continue' is not
skipped. Anyway, from what I can see it's not helping:
old ~ 2m
..*>****>*********>****>**>***>*>**>*>>**>*****>*****>******>*****>******>******>********>**>*****>***>***>***>*****************>*******>**>***>>~>**>*..
new ~ 30s
..*>*>*>*********>****>**>***>*>**>*>>**>*****>*.
It might look like it's saving time. It's about 1/3 of the way
through and it's only 30s, and it's missing lots of field setting.
I'm not sure why or how.