Explain the theory of what the problem is and why the output looks
like it does:
When a button navigates in the code above, this will press 'continue'
on the next page and look like it's continuing when it might be
clicking a button that continues. That will also give the wrong
impression about whether a variable was set.
Also once again raises the question of a continue button that sets
a variable getting 'set' before the fields on the page.