Get some boring selectors and their table vars, except danota
Also should be able to handle `_field_n`.

Architecture has morphed a lot and I'm not sure I love what it
looks like now. Also, no way to tell when something's base64, so
we don't really know when we need to decode further. Can there
be some kind of indicator.