Remove unused steps and then unused scope stuff, clean a bit.
These are steps I think nothing is using and/or are redundant now.
It removes a lot of steps that depend on visible text as opposed
to things that won't change when translated.
Also comments out checking value of checkbox for now as I think
it will need major changes and hasn't been used by anyone.
Closes #238, closes #240, closes #244 maybe
Addresses #227