Re: #138, Add variable information to reports
Address #138. Notes and caveats also covered in #138:
- End of table report is chronological so that people can copy/
paste it into their files if desired - alphabetical listing of
vars is recommended. (most useful for used vars)
- Problem with double char escaping (e.g. \\n) with Steps, but
avoids accidental new lines from story table strings.
- While running tests, page continuation symbol is printed in
during-test output, even if an error is about to be thrown for that
page. (problem?)
- All vars are displayed as table rows, even if a Step was used.
- Report shows original row, not the rows actually used in the
table. The concern is that it will be confusing otherwise, even
though it then gives less insight into what may have happened wrong
so we'll have to try to count on the warnings we've set up.
- The report is pretty complicated to read when multiple story
tables are used in one scenario.