Close #367, fix erroring too soon on no module error which
needs to give a chance for the module to reload. Also, added
another check to the setup pull as per instructions. See issue
comment for details. That should supposedly help this not happen,
but I did add a line in the report if it still happens and we'll
hopefully be able to see if that's not working and then report on
it.

Hope these are enough to fix this problem finally, but I don't know
of a way to test this.