Separate Scenario report heading vs. filename vs scenario id
Basically, instead of putting numbers in scenario headers and/or
adjusting tests to allow those to pass, I'm trying to make a more
extensive change to how we're managing that whole system.

Adds base-scenario-file-name creation and adds tests for that too.
Uses "-" as separator for major items as long as it doesn't
break our .gitignore and artifact-getting name matching in the
workflows. A change in that
will have to wait for a breaking version.