Add step to get page's vars and vals for v3. Unintegrated with the (#497)
other JSON step because I need to publish this one. It's already been
field tested by the user who requested it, so I'm going to forgo
a review in order to be able to move this forwards.

That is, the other Step is in a PR under review and I don't know
when someone will have time to take a look.

Addresses #454 for v3. Won't close till it's added to v4.