Add step to get page's vars and vals for v4. Unintegrated with the (#498)
other JSON Step because I need to publish this one and the other
is in an unreviewed PR and I don't know when someone will have
time to take a look. It's already been field tested by the user
who requested it, so I'm going to forgo a review in order to be
able to move this forwards.

Closes #454 for v4 when published.

Will wait till tests have passed.