Allow a developer to wait as a first Step for v3. Address #387. Add test. (#459)
This adds a temporary function for waiting in scope.js that will soon be replaced by a singular function in the utils folder. That said, that function is part of v4 changes, and this can improve v3 as well, so we might as well not wait.

Addresses #387, after a little confusion. Should add this to the testing docs as a 'first step' after this is merged.

[Also, generally adds safety measures for when page does not exist.]

------------------

* Allow a developer to wait as a first Step. Close #387. Add test.

* Offer proper fallback for wait options