Tap element2 (#556)
Allows you to tap an arbitrary element with a specific ID, or a bootstrap tab, made with the `tabbed_templates_html` function from ALToolbox.

If tapping arbitrary elements, you can specify a wait time ("and wait X seconds", or "and wait 1 second"). If tapping tabs, the framework listens to the `transitionend` event, waiting for the tab to fully transition before continuing. If a tab ID or element selector is mispelled or doesn't find any matching elements on the page, it fails and prints the message "Couldn't find ... on the page, is there a typo?"

Tests for this feature are in `interactive_steps.feature`, and they use the `test_taps.yml` in the ALAutomatedTestingTests repo.