Handle everything when object drop downs are added (#564)
* Handle everything when object drop downs are added

In https://github.com/plocket/docassemble-ALAutomatedTestingTests/pull/188

* Reset expected_status and expected_in_report

At beginning of each scenario, reset the variables expected_status and
expected_in_report of scope to `null`.

Otherwise, we maintain the `expected_in_report` is an empty array (`[]`) which
is truthy in Javascript. So after it's reset to [], we'll execute the code
checking to see if all of the specified lines (all 0 of them) are in the
report, which all of them will be and we will succeed every test.