Handle reload error for initial page load, abstract throwing server
reload errors.

Initial page load handling is untested because that's hard to do
locally. We may have to figure out some other way to test. Remote
testing for this probably won't work either as the timing would
be impossible - we'd have to start running setup on both tests
and hope that the server timeout didn't interfere with the other
setup, just with the page load once the tests had started.
Maybe make two local repos and set one up and then time the test
and new setup so they overlap? I'm not sure.