Discover and mend the immediate reload oversight which was
the "continue" Step using `wrapPromiseWithTimeout()` to ensure
preventing an infinite loop. It was causing a timeout too early
independently from the reload timeout. The problem now is where
else needs to deal with it, where to put the reoload complexity,
and how to make it clear where that's happening since it's so
far removed from a lot of the action. Also, is it only a problem
when elements are tapped, or other times as well? For example,
when downloading documents. Maybe that counts as tapping an
element.