Normalize a table that doesn't have the header row (#578)
* Throw an error if a story table doesn't have the header row

We require the header row to set the right attributes (`var`, `value`, and
`trigger`) to the row objects. If you don't have the header row, you'll get an
obscure error, usually that "no variables" matched. This gives a much clearer
error and can be resolved more quickly

* Allow tables to not include header row

Tries to parse tables using a header row first, but if there isn't one, then we
just take the raw columns and make objects out of them.

Fixes #257

Co-authored-by: plocket <52798256+plocket@users.noreply.github.com>