Close #392, avoid server reload errors more and handle them more gracefully (#570)
* Start tracking server status
* Creates untested way to wait for server to reload
* Confirm server reload check is triggering and working by
implementing roughly in story table step. A bit of a mess as I
need to keep confirming its working state as I go forward.
* Broken. Add waiting for server response in some places. See more...

Currently, the story table page id loop rejection doesn't error.
Waiting for server response also seems to not error on rejection.
Need to throw errors instead? This is probably why the try/catch
block wasn't working for triggering the check (earlier on)

Also, the puppeteer will timeout before the page id story table
loop times out, leaving the process unexited until the server
waiting and page id loop finish. How do we exit those immediately
when necessary? Or will it not be necessary when everything else
is lined up?

* Add throw in a couple places, make functions async some places
* Abstract throwing server reload error, add brainstorms
* Solve multiple promises being created, prep for custom value
* Clean up and add to comments
* Solve browser closing too early
* Handle reload error for initial page load, abstract throwing server reload errors.

Initial page load handling is untested because that's hard to do
locally. We may have to figure out some other way to test. Remote
testing for this probably won't work either as the timing would
be impossible - we'd have to start running setup on both tests
and hope that the server timeout didn't interfere with the other
setup, just with the page load once the tests had started.
Maybe make two local repos and set one up and then time the test
and new setup so they overlap? I'm not sure.

* Clarify message about server reloading
* Change server tracker from recursive to `while` loop
* Discover and mend the immediate reload oversight which was
the "continue" Step using `wrapPromiseWithTimeout()` to ensure
preventing an infinite loop. It was causing a timeout too early
independently from the reload timeout. The problem now is where
else needs to deal with it, where to put the reoload complexity,
and how to make it clear where that's happening since it's so
far removed from a lot of the action. Also, is it only a problem
when elements are tapped, or other times as well? For example,
when downloading documents. Maybe that counts as tapping an
element.
* Simplify timeout catching
* Fix server sign in not erroring properly on server reload
* Handle tapping element when reloading. I think...
Async is always hard to test. Basically, cleaned up where the
page was being closed and where the browser was being created.
Not sure if `sign_in` needs to create a browser now, or a new
page, etc, but I think it's enough that the cleanup is handled
consistently.
* Clean up old page closing code
* Pass tests for tapping var button and sign out at end of step
* Test sign out, continuing two ways, reload on both attempts,
and reload completes during first test.
* Comment out local-only tests
* Remove unnecessary test
* Simplify (I hope) creation of clickWait
* Clean up some cosmetics found in review
* Fix wrong var name used

Co-authored-by: Bryce Willey <Bryce.Steven.Willey@gmail.com>

* Ensure full time is waited between server reload checks

Co-authored-by: Bryce Willey <Bryce.Steven.Willey@gmail.com>