Bad commit: unit tests and features are broken
unit tests are broken because I added `multiple` as an attribute to the fields,
so now the deep equals are wrong, features are broken because we can't use
multipleselects, object or not. The entire framework of passing info to the
funnel needs to change so we can associate the decoded values from story tables
to the encoded values in the page.