Makes `...there_is_another | false |` okay in tables (#580)
* Makes `...there_is_another | false |` okay in tables

`there_is_another | true` is invalid in a story table, because it will always be
true, and start an infinite loop of adding more elements to whatever list is
gathering.

However, you might be asked to add elements to a pre-filled list. In my case, I
add Individuals to the `users` and `other_parties` from existing EFM case
information, and then ask if the interviewee wants to add more users. Therefore
I can't use `target_number` in the story table, as the number won't include the
number of elements already added to the list (which Kiln can't figure out). I
can however, use `there_is_another = False`, which will terminate.

* Add test for `there_is_another | False`

Ensure that no warnings are thrown in the report, and that it can successfully
pass.

Needed to add functions to check for the non-existance of things in the report,
and corrected some misspellings in other feature files that I came across.