Add a timestamp to project names instead of a number
Trying to deal with intermittent bug where alkiln runs tests on
what's supposed to be new code, but instead it's old code and test
results are wrong. We've seen this on one server at least. Our guess
is that something isn't being properly deleted on the test server,
but also that pulls aren't ovewritting it. Also, it goes away after
a couple days.