Use Axios's `toJSON` method to make shorter errors (#632)
* Use Axios's `toJSON` method to make shorter errors

On an example erroring docassemble server (for example, this server returning
a [502 to a delete project](https://github.com/SuffolkLITLab/docassemble-EFSPIntegration/actions/runs/3641406524/jobs/6147303425)),
the entire axios error will print, which is that case in 1.8k lines, most of it
being garbage.

Axios has a `toJSON()` method on errors that should reduce length of the errors
being shown, as it will not print the most annoying parts of the long errors,
like the internal Axios socket objects

Still need to test, as it's difficult to get a server to 502 on
demand.

Fix #605.

* Actually fail when we log the errors
* Move `toJSON()` calls to docassemble_api_REST
* Add better comment to `process.exitCode`

Co-authored-by: plocket <52798256+plocket@users.noreply.github.com>