Bump versions: node v18 & other actions to latest
Updates our node to v18, a long-term-service release,
which is supported until 2025.

Additionally, checkout, setup-node, and upload-artifact all ran on
node 12, which is reaching end of life, and won't be supported on
GitHub anymore. https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/

v3 of each of the above run with node 16.
* https://github.com/actions/checkout/blob/v3/action.yml#L78
* https://github.com/actions/setup-node/blob/v3/action.yml#L36
* https://github.com/actions/upload-artifact/blob/v3.0.0/action.yml#L27

Note: v16 does reach end of life in 2023/09. GitHub hasn't announced anything
about moving away from it yet, but judging on their timeline for deprecating
v12, I would guess that GitHub continue supporting  v16 until the end of 2024.

Fixes #645, #625, and #615.