Installable Kiln + Working on Windows! (#624)
Adds a new script: `alkiln-run`, which acts like `npm run cucumber`, but can
be run in any directory, not just in an npm package. Now, we no longer need to
create a false package.json in each docassemble package that we run in. To
do this, we upgrade cucumber to v8.6.0, and use the cucumber javascript API
to run tests.

Some additional changes:

* Make multiple tags work when passed to alkiln-run

Cucumber expects them in a boolean expression, i.e. `(@a0 or @a1) and @random`. Now, you can pass a list (with or without `@`) or use full boolean expressions.

* don't print the ["publish this cucumber report" message](https://github.com/cucumber/cucumber-js/blob/main/docs/configuration.md#options)

* Change action to install directly from the Github repo  v5 branch, for now, until v5 has a wider release.

---------

Co-authored-by: plocket <52798256+plocket@users.noreply.github.com>