Implement initial manually triggered language tests. This iteration
forces the author to use `Examples:` in every scenario, which is a bit burdensome. Working on a solution involving env vars and bash loops in the action, as well as another tag that will allow those kinds of tests to skip tests that should avoid testing with different languages (such as tests that are language-specific, like testing for different phrases).