Download the HTML for the page when screenshotting
IMO it's best to think of it as a part of the screenshot (enhanced screenshot).
We also replace relative URLs for CSS / JS with absolute URLs, adding the
server url in front of them. This isn't perfect, as CORS will prevent the
page from fully loading, but it's close enough, people will still have the
screenshot if they want exactly what the page looked like.