Pdfjs (#774)
Added the ability to compare PDFs. An example usage of the step is: `I expect the baseline PDF "something-in-sources.pdf" and the new PDF "downloaded.pdf" to be the same`. It compares both the text of the PDF, then the filled fields and their values.

Continues either if the PDF to compare couldn't be found (a developer error), or if the PDFs are different, since it doesn't affect interview progression.

Internal refactors:

* splits out the functionality to find files in the sources folder from `uploadFiles` to `findFiles`, which is used by both `uploadFiles` and `compare_pdfs`. 
* Also corrected a missed `scope.addToReport -> reports.addToReport` refactor.

---------

Co-authored-by: plocket <52798256+plocket@users.noreply.github.com>