Fix #725, partial filename match for downloading fails
The failure is actually during the comparison - the file was being
saved using the partial name the author gave in the Step when
it should have been using the file's downloaded name.

Note that our current download method will, in ideal cases, let an
author re-name the downloaded file. Since it might not be
consistent, we might still avoid implementing this at this point.
Also, in ideal cases, the Step does already continue without
waiting for the whole file to download, which addresses #492. That
happened a while ago. It's worth considering closing that issue at
this point. We have to research under what circumstances a file
would fail to download with a `fetch` while still being able to
download with a click. Maybe with an encrypted interview.

Co-authored-by: ethanstrominger <ethanstrominger2@gmail.com>